apps/telegram-forwarder/requirements.txt
telethon>=1.28.0
python-dotenv>=1.0.0
cryptg>=0.4.0  # For faster downloads/uploads
